requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
deepface>=0.0.80
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0
